+++ apps/website/package.json
@@ -19,13 +19,13 @@
   "scripts": {
     "build:next": "next build --webpack",
     "build:next-sitemap": "next-sitemap --config next-sitemap.config.js",
-    "build:push-sw": "tsup --config sw.tsup.config.ts",
-    "build:trpc-types": "tsup --config trpc.tsup.config.ts",
+    "build:push-sw": "tsdown --config sw.tsdown.config.ts",
+    "build:trpc-types": "tsdown --config trpc.tsdown.config.ts",
     "build:manifest": "tsx ./build-scripts/create-manifest.ts",
     "build:presets": "tsx ./build-scripts/update-presets.ts",
     "dev:next": "next dev --webpack",
-    "dev:push-sw": "tsup --watch --config sw.tsup.config.ts",
-    "dev:trpc-types": "tsup --watch --config trpc.tsup.config.ts",
+    "dev:push-sw": "tsdown --watch --config sw.tsdown.config.ts",
+    "dev:trpc-types": "tsdown --watch --config trpc.tsdown.config.ts",
     "build": "pnpm run --sequential \"/^build:(?!presets$).*$/\"",
     "dev": "pnpm run \"/^(build:manifest|dev:.*)/\"",
     "postinstall": "pnpm build:trpc-types",
@@ -134,7 +134,7 @@
     "prettier-plugin-tailwindcss": "^0.7.2",
     "tailwindcss": "^4.2.2",
     "ts-morph": "^27.0.2",
-    "tsup": "^8.5.1",
+    "tsdown": "^0.21.9",
     "tsx": "^4.21.0",
     "typescript-eslint": "^8.58.1",
     "vitest": "^4.1.4",
