--- apps/website/trpc.tsup.config.ts
@@ -1,18 +1,0 @@
-import { defineConfig } from "tsup";
-
-const config = defineConfig({
-  entry: {
-    trpc: "src/server/trpc/router/_app.ts",
-  },
-  outDir: "dist",
-  format: "esm",
-  dts: {
-    only: true,
-    compilerOptions: {
-      incremental: false,
-    },
-  },
-  clean: false,
-});
-
-export default config;
